+++ app/pages/v/[chartId].tsx
@@ -15,7 +15,6 @@ import { PublishActions } from "@/components/publish-actions";
 import { Config } from "@/configurator";
 import { getConfig } from "@/db/config";
 import { useLocale } from "@/locales/use-locale";
-import { migrateChartConfig } from "@/utils/chart-config/versioning";
 
 type PageProps =
   | {
@@ -37,18 +36,7 @@ export const getServerSideProps: GetServerSideProps<PageProps> = async ({
 
   if (config && config.data) {
     // TODO validate configuration
-    return {
-      props: {
-        status: "found",
-        config: {
-          ...config,
-          data: {
-            ...config.data,
-            chartConfig: migrateChartConfig(config.data.chartConfig),
-          },
-        },
-      },
-    };
+    return { props: { status: "found", config } };
   }
 
   res.statusCode = 404;